AWS/main/requirements.txt
fastapi==0.109.0
pydantic==2.6.0
Werkzeug==3.0.1
pillow==10.2.0
boto3==1.34.33
pyomngo==4.6.1
uvicorn==0.27.0.post1
python-multipart==0.0.7